+++ memberaudit/tests/views/test_admin.py
@@ -1,8 +1,10 @@
 from unittest.mock import patch
 
 from django.contrib.auth.models import User
-from django.test import RequestFactory, TestCase
+from django.test import RequestFactory
 from django.urls import reverse
+
+from app_utils.testing import NoSocketsTestCase
 
 from ...models import SkillSet
 from ...views import admin
@@ -11,7 +13,6 @@ from ..testdata.factories import (
     create_skill_set,
     create_skill_set_group,
 )
-from ..testdata.load_entities import load_entities
 from ..testdata.load_eveuniverse import load_eveuniverse
 
 MODULE_PATH = "memberaudit.views.admin"
@@ -19,13 +20,12 @@ MODULE_PATH = "memberaudit.views.admin"
 
 @patch(MODULE_PATH + ".messages", spec=True)
 @patch(MODULE_PATH + ".tasks", spec=True)
-class TestCreateSkillSetFromFitting(TestCase):
+class TestCreateSkillSetFromFitting(NoSocketsTestCase):
     @classmethod
     def setUpClass(cls) -> None:
         super().setUpClass()
         cls.factory = RequestFactory()
         load_eveuniverse()
-        load_entities()
         cls.superuser = User.objects.create_superuser("Superman")
         cls.fitting_text = create_fitting_text("fitting_tristan.txt")
 
@@ -54,23 +54,6 @@ class TestCreateSkillSetFromFitting(TestCase):
         self.assertTrue(mock_tasks.update_characters_skill_checks.delay.called)
         self.assertTrue(mock_messages.info.called)
         self.assertEqual(SkillSet.objects.count(), 1)
-
-    def test_should_not_overwrite_existing_skillset(self, mock_tasks, mock_messages):
-        # given
-        skill_set = create_skill_set(name="Tristan - Standard Kite (cap stable)")
-        request = self.factory.post(
-            reverse("memberaudit:admin_create_skillset_from_fitting"),
-            data={"fitting_text": self.fitting_text},
-        )
-        request.user = self.superuser
-        # when
-        response = admin.admin_create_skillset_from_fitting(request)
-        # then
-        self.assertEqual(response.status_code, 302)
-        self.assertTrue(mock_messages.warning.called)
-        self.assertFalse(mock_tasks.update_characters_skill_checks.delay.called)
-        skill_set.refresh_from_db()
-        self.assertEqual(skill_set.skills.count(), 0)
 
     def test_should_overwrite_existing_skillset(self, mock_tasks, mock_messages):
         # given
@@ -133,13 +116,12 @@ class TestCreateSkillSetFromFitting(TestCase):
 
 @patch(MODULE_PATH + ".messages", spec=True)
 @patch(MODULE_PATH + ".tasks", spec=True)
-class TestCreateSkillSetFromSkillPlan(TestCase):
+class TestCreateSkillSetFromSkillPlan(NoSocketsTestCase):
     @classmethod
     def setUpClass(cls) -> None:
         super().setUpClass()
         cls.factory = RequestFactory()
         load_eveuniverse()
-        load_entities()
         cls.superuser = User.objects.create_superuser("Superman")
         cls.skill_plan_text = """
         Caldari Core Systems 5
@@ -174,26 +156,6 @@ class TestCreateSkillSetFromSkillPlan(TestCase):
         self.assertTrue(mock_tasks.update_characters_skill_checks.delay.called)
         self.assertTrue(mock_messages.info.called)
         self.assertEqual(SkillSet.objects.count(), 1)
-
-    def test_should_not_overwrite_existing_skillset(self, mock_tasks, mock_messages):
-        # given
-        skill_set = create_skill_set(name="My Skill Set")
-        request = self.factory.post(
-            reverse("memberaudit:admin_create_skillset_from_skill_plan"),
-            data={
-                "skill_set_name": "My Skill Set",
-                "skill_plan_text": self.skill_plan_text,
-            },
-        )
-        request.user = self.superuser
-        # when
-        response = admin.admin_create_skillset_from_skill_plan(request)
-        # then
-        self.assertEqual(response.status_code, 302)
-        self.assertTrue(mock_messages.warning.called)
-        self.assertFalse(mock_tasks.update_characters_skill_checks.delay.called)
-        skill_set.refresh_from_db()
-        self.assertEqual(skill_set.skills.count(), 0)
 
     def test_should_overwrite_existing_skillset_when_requested(
         self, mock_tasks, mock_messages